+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 FEATURES:
 
+* Experimental Dynamodb as persistent datastore
 * Centralise utilities helper and data structures and algorithms
 
 ## v0.1.22-23 -  12/12/2023